# server/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MediaTracker",
        "version": "0.0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "security": [
        {
            "ApiKey": []
        },
        {
            "ApiHeader": []
        }
    ],
    "paths": {
...
        "/api/items/paginated": {
            "get": {
                "description": "Get items",
                "tags": ["Items"],
                "operationId": "paginated",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/MediaItemOrderBy"
                                }
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/SortOrder"
                                }
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyOnWatchlist",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlySeenItems",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithNextEpisodesToWatch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithNextAiring",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithUserRating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithoutUserRating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "nullable": true
                        }
                    },
                    {
                        "name": "mediaType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/MediaType"
                                }
                            ],
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MediaItemItemsResponse"
                                            }
                                        },
                                        "page": {
                                            "type": "number"
                                        },
                                        "totalPages": {
                                            "type": "number"
                                        },
                                        "from": {
                                            "type": "number"
                                        },
                                        "to": {
                                            "type": "number"
                                        },
                                        "total": {
                                            "type": "number"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "from",
                                        "page",
                                        "to",
                                        "total",
                                        "totalPages"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items": {
            "get": {
                "description": "Get items",
                "tags": ["Items"],
                "operationId": "get",
                "parameters": [
                    {
                        "name": "mediaType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/MediaType"
                                }
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/MediaItemOrderBy"
                                }
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/SortOrder"
                                }
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyOnWatchlist",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlySeenItems",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithNextEpisodesToWatch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithNextAiring",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithUserRating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "onlyWithoutUserRating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediaItemItemsResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "in": "query",
                "name": "token"
            },
            "ApiHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "Access-Token"
            }
        },
        "schemas": {
            "TvEpisode": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "nullable": true
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "episodeNumber": {
                        "type": "number"
                    },
                    "seasonNumber": {
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string",
                        "nullable": true
                    },
                    "tvShowId": {
                        "type": "number",
                        "nullable": true
                    },
                    "seasonId": {
                        "type": "number",
                        "nullable": true
                    },
                    "tmdbId": {
                        "type": "number",
                        "nullable": true
                    },
                    "imdbId": {
                        "type": "string",
                        "nullable": true
                    },
                    "runtime": {
                        "type": "number",
                        "nullable": true
                    },
                    "seenHistory": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Seen"
                        },
                        "nullable": true
                    },
                    "userRating": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/UserRating"
                            }
                        ],
                        "nullable": true
                    },
                    "lastSeenAt": {
                        "type": "number",
                        "nullable": true
                    },
                    "seasonAndEpisodeNumber": {
                        "type": "number",
                        "nullable": true
                    },
                    "seen": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "tvShow": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/MediaItemItemsResponse"
                            }
                        ],
                        "nullable": true
                    },
                    "isSpecialEpisode": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "episodeNumber",
                    "isSpecialEpisode",
                    "seasonNumber",
                    "title"
                ]
            },
            "Seen": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "nullable": true
                    },
                    "date": {
                        "type": "number"
                    },
                    "mediaItemId": {
                        "type": "number"
                    },
                    "seasonId": {
                        "type": "number",
                        "nullable": true
                    },
                    "episodeId": {
                        "type": "number",
                        "nullable": true
                    },
                    "userId": {
                        "type": "number"
                    }
                },
                "required": ["date", "mediaItemId", "userId"]
            },
            "UserRating": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "nullable": true
                    },
                    "mediaItemId": {
                        "type": "number"
                    },
                    "date": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    },
                    "rating": {
                        "type": "number",
                        "nullable": true
                    },
                    "review": {
                        "type": "string",
                        "nullable": true
                    },
                    "episodeId": {
                        "type": "number",
                        "nullable": true
                    },
                    "seasonId": {
                        "type": "number",
                        "nullable": true
                    }
                },
                "required": ["date", "mediaItemId", "userId"]
            },
            "MediaItemItemsResponse": {
                "allOf": [
                    {
                        "type": "object",
                        "properties": {
                            "tmdbId": {
                                "type": "number",
                                "nullable": true
                            },
                            "imdbId": {
                                "type": "string",
                                "nullable": true
                            },
                            "tvmazeId": {
                                "type": "number",
                                "nullable": true
                            },
                            "igdbId": {
                                "type": "number",
                                "nullable": true
                            },
                            "openlibraryId": {
                                "type": "string",
                                "nullable": true
                            },
                            "audibleId": {
                                "type": "string",
                                "nullable": true
                            },
                            "id": {
                                "type": "number",
                                "nullable": true
                            },
                            "numberOfSeasons": {
                                "type": "number",
                                "nullable": true
                            },
                            "status": {
                                "type": "string",
                                "nullable": true
                            },
                            "platform": {
                                "type": "string",
                                "nullable": true
                            },
                            "title": {
                                "type": "string"
                            },
                            "originalTitle": {
                                "type": "string",
                                "nullable": true
                            },
                            "poster": {
                                "type": "string",
                                "nullable": true
                            },
                            "backdrop": {
                                "type": "string",
                                "nullable": true
                            },
                            "tmdbRating": {
                                "type": "number",
                                "nullable": true
                            },
                            "releaseDate": {
                                "type": "string",
                                "nullable": true
                            },
                            "overview": {
                                "type": "string",
                                "nullable": true
                            },
                            "lastTimeUpdated": {
                                "type": "number"
                            },
                            "source": {
                                "type": "string"
                            },
                            "network": {
                                "type": "string",
                                "nullable": true
                            },
                            "url": {
                                "type": "string",
                                "nullable": true
                            },
                            "runtime": {
                                "type": "number",
                                "nullable": true
                            },
                            "mediaType": {
                                "$ref": "#/components/schemas/MediaType"
                            },
                            "genres": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "nullable": true
                            },
                            "numberOfEpisodes": {
                                "type": "number",
                                "nullable": true
                            },
                            "developer": {
                                "type": "string",
                                "nullable": true
                            },
                            "authors": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "nullable": true
                            },
                            "narrators": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "nullable": true
                            },
                            "language": {
                                "type": "string",
                                "nullable": true
                            },
                            "needsDetails": {
                                "type": "boolean",
                                "nullable": true
                            }
                        },
                        "required": [
                            "lastTimeUpdated",
                            "mediaType",
                            "source",
                            "title"
                        ]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "isSearchResult": {
                                "type": "boolean",
                                "nullable": true
                            },
                            "hasDetails": {
                                "type": "boolean",
                                "nullable": true
                            },
                            "poster": {
                                "type": "string",
                                "nullable": true
                            },
                            "posterSmall": {
                                "type": "string",
                                "nullable": true
                            },
                            "backdrop": {
                                "type": "string",
                                "nullable": true
                            },
                            "seenHistory": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Seen"
                                },
                                "nullable": true
                            },
                            "unseenEpisodesCount": {
                                "type": "number",
                                "nullable": true
                            },
                            "userRating": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserRating"
                                    }
                                ],
                                "nullable": true
                            },
                            "upcomingEpisode": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/TvEpisode"
                                    }
                                ],
                                "nullable": true
                            },
                            "nextAiring": {
                                "type": "string",
                                "nullable": true
                            },
                            "onWatchlist": {
                                "type": "boolean",
                                "nullable": true
                            },
                            "lastSeenAt": {
                                "type": "number",
                                "nullable": true
                            },
                            "seen": {
                                "type": "boolean",
                                "nullable": true
                            },
                            "firstUnwatchedEpisode": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/TvEpisode"
                                    }
                                ],
                                "nullable": true
                            }
                        }
                    }
                ]
            },
            "MediaType": {
                "enum": ["audiobook", "book", "movie", "tv", "video_game"],
                "type": "string"
            },
...
            "MediaItemOrderBy": {
                "enum": [
                    "lastSeen",
                    "mediaType",
                    "nextAiring",
                    "releaseDate",
                    "status",
                    "title",
                    "unseenEpisodes"
                ],
                "type": "string"
            },
            "SortOrder": {
                "enum": ["asc", "desc"],
                "type": "string"
            },
...
        }
    }
}